Ancestor tree I
decided for a tree view of the ancestors. Sunburst is postponed.

Recursiv call of database implemented to get the fathers ancestors .
Correct handling of jdbc statement needs still  to be done -- especially
closure of statements and resultsets.

Recursiv call needs all to be done for the mother on the server side .

on Client side . New menu available on the right side, when a person is
retrieved.
A Dialog window popsup when the ancestor tree view is selected , showing
the ancestors. Perhaps the Dialog should be changed to a complete new
browser window